requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
kafka-python>=2.0.2
pandas>=2.0.0
psycopg2-binary>=2.9.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
tabulate>=0.9.0
python-dateutil>=2.8.2
circuitbreaker>=1.4.0
prometheus-client>=0.17.0
python-json-logger>=2.0.7
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
pytest>=7.4.0
pytest-asyncio>=0.21.0 